v3.1.0

+++ src/contextmenu/delete.ts
@@ -1,22 +1,29 @@
-import { ChannelType } from 'discord.js';
 import { readFileSync } from 'fs';
 import { respond } from '../utils/helperFunctions';
 import userModel from '../models/userModel';
-import { ContextMenuCommand, WebhookMessages } from '../typedef';
+import { canManageWebhooks, missingPermissions } from '../utils/permissionHandler';
+import { WebhookMessages } from '../typings/data/general';
+import { ContextMenuCommand } from '../typings/handle';
+import { isInGuild } from '../utils/checkUserState';
 
-const name: ContextMenuCommand['name'] = 'Delete ❌';
 export const command: ContextMenuCommand = {
-	name: name,
 	data: {
-		name: name,
+		name: 'Delete ❌',
 		type: 3,
 		dm_permission: false,
 	},
-	sendCommand: async (client, interaction) => {
+	sendCommand: async (interaction) => {
+
+		/* This shouldn't happen as dm_permission is false. */
+		if (!isInGuild(interaction)) { return; }
+
+		if (await missingPermissions(interaction, [
+			'ManageWebhooks', // Needed for webhook interaction
+		]) === true) { return; }
 
 		/* This gets the webhookCache and userData */
 		const webhookCache = JSON.parse(readFileSync('./database/webhookCache.json', 'utf-8')) as WebhookMessages;
-		const userData = await userModel.findOne(u => u.userId.includes(webhookCache[interaction.targetId]?.split('_')[0] || '')).catch(() => { return null; });
+		const userData = userModel.find(u => u.userId.includes(webhookCache[interaction.targetId]?.split('_')[0] || ''))[0] ?? null;
 
 		/* This is checking if the user who is trying to delete the message is the same user who sent the message. */
 		if (userData === null || !userData.userId.includes(interaction.user.id)) {
@@ -35,23 +42,9 @@ export const command: ContextMenuCommand = {
 		if (interaction.channel === null) { throw new Error('Interaction channel is null.'); }
 		const webhookChannel = interaction.channel.isThread() ? interaction.channel.parent : interaction.channel;
 		if (webhookChannel === null) { throw new Error('Webhook can\'t be edited, interaction channel is thread and parent channel cannot be found'); }
-		if (webhookChannel.type === ChannelType.DM) { throw new Error('Webhook can\'t be edited, channel is DMChannel.'); }
-		const webhook = (await webhookChannel
-			.fetchWebhooks()
-			.catch(async (error) => {
-				if (error.httpStatus === 403) {
-					await interaction.channel?.send({ content: 'Please give me permission to create webhooks 😣' }).catch((err) => { throw err; });
-				}
-				throw error;
-			})
-		).find(webhook => webhook.name === 'PnP Profile Webhook') || await webhookChannel
-			.createWebhook({ name: 'PnP Profile Webhook' })
-			.catch(async (error) => {
-				if (error.httpStatus === 403) {
-					await interaction.channel?.send({ content: 'Please give me permission to create webhooks 😣' }).catch((err) => { throw err; });
-				}
-				throw error;
-			});
+		if (await canManageWebhooks(interaction.channel) === false) { return; }
+		const webhook = (await webhookChannel.fetchWebhooks()).find(webhook => webhook.name === 'PnP Profile Webhook')
+			|| await webhookChannel.createWebhook({ name: 'PnP Profile Webhook' });
 
 		/* Deleting the message. */
 		await webhook.deleteMessage(interaction.targetId, interaction.channel.isThread() ? interaction.channel.id : undefined);
